:racehorse: Speed things up a little bit
While I liked the idea of visiting every single "pixel" in the "canvas" each
time, one pixel per animation loop, it did get a bit boring as the set
started to take shape. So this changes things up a bit so that on each walk
through the canvas we only visibly visit the pixels that are left.

And once we've exhausted all of the available steps, we don't do any work
during each animation loop.