:bug: Always render the scene every animation loop
It seemed like a good idea at the time, and in the web page it made no
obvious difference, but *not* rendering the scene every animation loop means
that once there's no changes made any more the scene doesn't work well in
VR. It doesn't redraw based on your head movement or body movement, and the
controllers don't get updated as you move your hands, etc.

Oops!

So this fixes that.